+++ python/GafferOSLUI/OSLShaderUI.py
@@ -40,6 +40,7 @@ import IECore
 
 import Gaffer
 import GafferUI
+import GafferSceneUI
 
 import GafferOSL
 
@@ -177,12 +178,21 @@ def __plugNoduleLabel( plug ) :
 
 def __plugActivator( plug ) :
 
+	# Prefer our own format, in which activation is defined by a single
+	# OSL expression with access to all plug values and all OSL constructs.
+
 	node = plug.node()
 	expression = node.parameterMetadata( plug, "enabledExpression" )
 	if expression :
 		return node.evaluateActivatorExpression( expression )
 
-	return True
+	# Fall back to a far less satisfying but more pervasive format.
+
+	node = plug.node()
+	return not GafferSceneUI.ShaderUI._evaluateConditionalLock(
+		node["parameters"],
+		lambda key : node.parameterMetadata( plug, key )
+	)
 
 def __plugVisibilityActivator( plug ) :
 
@@ -191,7 +201,11 @@ def __plugVisibilityActivator( plug ) :
 	if expression :
 		return node.evaluateActivatorExpression( expression )
 
-	return True
+	node = plug.node()
+	return GafferSceneUI.ShaderUI._evaluateConditionalVisibility(
+		node["parameters"],
+		lambda key : node.parameterMetadata( plug, key ),
+	)
 
 Gaffer.Metadata.registerNode(